// be/types/logic/orders/insert/duplicate.d.ts
export type filter_object = {
order_id: string,
order_date: string,
order_status: string,
order_detail_id: string,
order_options: string, 
order_option_ids: string,
cust_id: string, 
full_name: string,
phone_number: string,
product_name: string,
product_code: string,
product_qt: string, 
notes: string, 
pallet: string,
tack: string, 
assembled: string, 
completed: string
ship_country: string,
ship_state: string, 
ship_city: string, 
ship_address: string, 
ship_postal_code: string,
tracking: string,
payment_amount: string, 
payment_method_id: string,
}